requirements.txt
streamlit
akshare
pandas
numpy
httpx
plotly
python-dotenv
mcp